jks/sun_crypto.py: Moved jks.sun_crypto._strip_pkcs5_padding into jks.util.strip_pkcs5_padding to prepare for use by other algorithms than JKS/JCE
